--- shared/packages/expectationManager/src/__tests__/test.spec.ts
@@ -1,7 +1,0 @@
-describe('tmp', () => {
-	test('tmp', () => {
-		// Note: To enable tests in this package, ensure that the "test" script is present in package.json
-		expect(1).toEqual(1)
-	})
-})
-export {}
