+++ shared/packages/expectationManager/src/lib/__tests__/trackedExpectation.spec.ts
@@ -1,9 +1,8 @@
 import { Expectation } from '@sofie-package-manager/api'
-import { getDefaultConstants } from '../lib/constants'
-import { getDefaultTrackedExpectation, sortTrackedExpectations } from '../lib/expectations'
-import { TrackedExpectation } from '../expectationTracker'
+import { getDefaultConstants } from '../constants'
+import { TrackedExpectation, getDefaultTrackedExpectation, sortTrackedExpectations } from '../trackedExpectation'
 
-test('sortExpectations', () => {
+test('sortTrackedExpectations', () => {
 	// ensure that sort order is correct
 
 	const exp: Expectation.FileCopy = {
